add vigenere cipher
Uploaded a program that performs the vigenere cipher on a given string. It requires a string (s) and a substring (sub) to encode the string by adding the index of each character in the alphabet. After adding the indexes, a new char is found by grabbing the char in the calculated position. The "fill()" function requires a string and an integer to make the string as long as the integer by removing a char if it is too long and adding a char if it is too short.